(VDB-637) Update contracts for 0.2.6
- Update addresses, ABI, and deployment blocks
- Change type of bite flip to reflect latest ABI

+++ transformers/events/bite/repository.go
@@ -72,7 +72,7 @@ func (repository BiteRepository) Create(headerID int64, models []interface{}) er
 
 		_, execErr := tx.Exec(
 			`INSERT into maker.bite (header_id, urn_id, ink, art, tab, flip, log_idx, tx_idx, raw_log)
-        			VALUES($1, $2, $3::NUMERIC, $4::NUMERIC, $5::NUMERIC, $6::NUMERIC, $7, $8, $9)
+					VALUES($1, $2, $3::NUMERIC, $4::NUMERIC, $5::NUMERIC, $6, $7, $8, $9)
 					ON CONFLICT (header_id, tx_idx, log_idx) DO UPDATE SET urn_id = $2, ink = $3, art = $4, tab = $5, flip = $6, raw_log = $9;`,
 			headerID, urnID, biteModel.Ink, biteModel.Art, biteModel.Tab, biteModel.Flip, biteModel.LogIndex, biteModel.TransactionIndex, biteModel.Raw,
 		)
